Change raw acceleration format
I'd previously assumed that it would be safe to assume consecutive raw
acceleration values are never more than 255 milliseconds apart. After
seeing some of the data, I'm not so sure anymore.

Change the data format to be able to write a full timestamp instead of
just the difference to the previous one, but only if the difference
doesn't fit into a uint8.