Hack: disable all but raw acceleration
Sending raw acceleration works now. However, with everything going on on
the device, the CPU seems to be pretty busy, and we're only getting a
sample every 90-or-so milliseconds, with occasional gaps of about 200ms.

If we disable everything else, we get samples 2-3ms apart. Everything is
just commented out for testing.